# backend/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /thread/:id/like:
    delete:
      consumes:
      - application/json
      description: Unlikes a thread
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Thread successfully unliked
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unlikes a thread
      tags:
      - threads
    post:
      consumes:
      - application/json
      description: Likes a thread
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Thread successfully liked
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Likes a thread
      tags:
      - threads
  /thread/:id/report:
    put:
      consumes:
      - application/json
      description: Reports a thread
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Thread successfully reported
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reports a thread
      tags:
      - threads
...
swagger: "2.0"